requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
livekit-api==0.7.1
pydantic==2.9.2
